riepilogo1/requirements.txt
flask
requests
google-cloud-storage
google-cloud-firestore
